requirements.txt
python-telegram-bot==13.15
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.8.5
pytz==2023.3
python-dotenv==1.0.0
apscheduler==3.10.1
logging==0.4.9.6
colorlog==6.7.0
pandas==2.0.3
numpy==1.24.4
tqdm==4.66.1
retry==0.9.2 